+++ src/components/_common/animate-height/AnimateHeight.module.scss
@@ -1,12 +1,12 @@
 .animated {
     overflow: hidden;
-    transition: height, opacity 100ms linear;
+    transition-property: height, opacity;
+    transition-timing-function: ease-in;
 }
 
 .fade {
     opacity: 100%;
     animation-name: fade;
-    animation-duration: 200ms;
     animation-fill-mode: forwards;
     animation-direction: alternate;
     animation-timing-function: linear;
@@ -19,6 +19,6 @@
     }
 
     to {
-        opacity: 0;
+        opacity: 25%;
     }
 }
